Complete index sub module

+++ web/src/main/java/com/order/bachlinh/web/component/search/index/spi/IndexWriter.java
@@ -1,5 +1,5 @@
 package com.order.bachlinh.web.component.search.index.spi;
 
 public interface IndexWriter {
-    int write(String keyword);
+    long write(String keyword);
 }
